Make "active-active" a valid production_type value (#249)
The active-active mode used to be identified by the vm_node_count parameter. This breaks possible workflows when someone wants to modify the node_count to a value lower than 2 for maintenance or other reasons.

This adds a non-breaking change to decouple the active-active mode from the vm_node_count parameter.